+++ subsys/lorawan/services/CMakeLists.txt
@@ -2,6 +2,8 @@
 
 if(CONFIG_LORAWAN_SERVICES)
 
+  zephyr_library()
+
   zephyr_library_sources(lorawan_services.c)
 
   zephyr_library_sources_ifdef(
@@ -17,12 +19,17 @@ if(CONFIG_LORAWAN_SERVICES)
 
   zephyr_library_sources_ifdef(
     CONFIG_LORAWAN_FRAG_TRANSPORT_DECODER_LOWMEM
-          frag_decoder_lowmem.c
+    frag_decoder_lowmem.c
   )
 
-  zephyr_library_sources_ifdef(
-    CONFIG_LORAWAN_REMOTE_MULTICAST
-    multicast.c
-  )
+  if(CONFIG_LORAWAN_FRAG_TRANSPORT_DECODER_SEMTECH)
+    zephyr_library_include_directories(
+      ${ZEPHYR_LORAMAC_NODE_MODULE_DIR}/src/apps/LoRaMac/common/LmHandler/packages
+    )
+  endif()
 
 endif()
+
+# The multicast service is loramac-node specific
+set(ZEPHYR_CURRENT_LIBRARY loramac-node)
+zephyr_library_sources_ifdef(CONFIG_LORAWAN_REMOTE_MULTICAST multicast.c)
