+++ subsys/lorawan/services/frag_decoder_lowmem.h
@@ -18,8 +18,9 @@
 #define FRAG_MAX_NB                                                                                \
 	(CONFIG_LORAWAN_FRAG_TRANSPORT_IMAGE_SIZE / CONFIG_LORAWAN_FRAG_TRANSPORT_MIN_FRAG_SIZE +  \
 	 1U)
-#define FRAG_MAX_SIZE  (CONFIG_LORAWAN_FRAG_TRANSPORT_MAX_FRAG_SIZE)
-#define FRAG_TOLERANCE (FRAG_MAX_NB * CONFIG_LORAWAN_FRAG_TRANSPORT_MAX_REDUNDANCY / 100U)
+#define FRAG_MAX_SIZE       (CONFIG_LORAWAN_FRAG_TRANSPORT_MAX_FRAG_SIZE)
+#define FRAG_MAX_REDUNDANCY (FRAG_MAX_NB * CONFIG_LORAWAN_FRAG_TRANSPORT_MAX_REDUNDANCY / 100U)
+#define FRAG_TOLERANCE      FRAG_MAX_REDUNDANCY
 
 #define FRAG_DEC_ONGOING                  (-1)
 #define FRAG_DEC_ERR_INVALID_FRAME        (-2)
